Update package description to include projectiles + doc comment edits
