+++ projectile.go
@@ -1,21 +1,27 @@
-// Simple physics projectile motion.
+package harmonica
+
+// This file defines simple physics projectile motion.
 //
 // Example usage:
 //
 //    // Run once to initialize.
-//    projectile := NewProjectile(FPS(60), Point{6.0, 100.0, 0.0}, Vector{2.0, 0.0, 0.0}, Vector{2.0, -9.81, 0.0})
+//    projectile := NewProjectile(
+//        FPS(60),
+//        Point{6.0, 100.0, 0.0},
+//        Vector{2.0, 0.0, 0.0},
+//        Vector{2.0, -9.81, 0.0},
+//    )
 //
 //    // Update on every frame.
 //    someUpdateLoop(func() {
-//      pos := projectile.Update()
+//        pos := projectile.Update()
 //    })
 //
 // For background on projectile motion see:
 // https://en.wikipedia.org/wiki/Projectile_motion
-package harmonica
 
-// Projectile is the representation of a projectile that has a position on a
-// plane and an acceleration and velocity
+// Projectile is the representation of a projectile that has a position on
+// a plane, an acceleration, and velocity.
 type Projectile struct {
 	pos       Point
 	vel       Vector
@@ -23,16 +29,16 @@ type Projectile struct {
 	deltaTime float64
 }
 
-// Point is a representation of a point which contains the X, Y, Z coordinates
-// of the point on a plane.
+// Point represents a point containing the X, Y, Z coordinates of the point on
+// a plane.
 type Point struct {
 	X, Y, Z float64
 }
 
-// Vector is a representation of a vector which carries a magnitude and a
-// direction.  We represent the vector as a point from the origin (0, 0) where
-// the magnitude is the euclidean distance from the origin and the direction is
-// the direction to the point from the origin.
+// Vector represents a vector carrying a magnitude and a direction. We
+// represent the vector as a point from the origin (0, 0) where the magnitude
+// is the euclidean distance from the origin and the direction is the direction
+// to the point from the origin.
 type Vector struct {
 	X, Y, Z float64
 }
@@ -48,12 +54,13 @@ type Vector struct {
 // (i.e. origin is located in the bottom-left corner)
 var Gravity = Vector{0, -9.81, 0}
 
-// TerminalGravity is a utility vector that represents gravity where
-// the coordinate plane's origin is on the top-right corner
+// TerminalGravity is a utility vector that represents gravity where the
+// coordinate plane's origin is on the top-right corner
 var TerminalGravity = Vector{0, 9.81, 0}
 
-// NewProjectile accepts a frame rate, and initial values for position, velocity, and acceleration and
-// returns a new projectile.
+// NewProjectile creates a new projectile. It accepts a frame rate and initial
+// values for position, velocity, and acceleration. It returns a new
+// projectile.
 func NewProjectile(deltaTime float64, initialPosition Point, initialVelocity, initalAcceleration Vector) *Projectile {
 	return &Projectile{
 		pos:       initialPosition,
@@ -77,7 +84,7 @@ func (p *Projectile) Update() Point {
 	return p.pos
 }
 
-// Position returns the position of the projectile
+// Position returns the position of the projectile.
 func (p *Projectile) Position() Point {
 	return p.pos
 }
